+++ xr.js
@@ -1,19 +1,25 @@
-(function (factory) {
-  if (typeof define === "function" && define.amd) {
-    define(["exports", "module"], factory);
-  } else if (typeof exports !== "undefined" && typeof module !== "undefined") {
+(function (global, factory) {
+  if (typeof define === 'function' && define.amd) {
+    define(['exports', 'module'], factory);
+  } else if (typeof exports !== 'undefined' && typeof module !== 'undefined') {
     factory(exports, module);
+  } else {
+    var mod = {
+      exports: {}
+    };
+    factory(mod.exports, mod);
+    global.xr = mod.exports;
   }
-})(function (exports, module) {
+})(this, function (exports, module) {
   /**
    * xr (c) James Cleveland 2015
    * URL: https://github.com/radiosilence/xr
    * License: BSD
    */
 
-  "use strict";
+  'use strict';
 
-  var res = function (xhr) {
+  var res = function res(xhr) {
     return {
       status: xhr.status,
       response: xhr.response,
@@ -33,67 +39,67 @@
     return t;
   };
 
-  var getParams = function (data, url) {
+  var getParams = function getParams(data, url) {
     var ret = [];
     for (var k in data) {
-      ret.push("" + encodeURIComponent(k) + "=" + encodeURIComponent(data[k]));
-    }if (url && url.split("?").length > 1) ret.push(url.split("?")[1]);
-    return ret.join("&");
+      ret.push('' + encodeURIComponent(k) + '=' + encodeURIComponent(data[k]));
+    }if (url && url.split('?').length > 1) ret.push(url.split('?')[1]);
+    return ret.join('&');
   };
 
   var Methods = {
-    GET: "GET",
-    POST: "POST",
-    PUT: "PUT",
-    DELETE: "DELETE",
-    PATCH: "PATCH",
-    OPTIONS: "OPTIONS"
+    GET: 'GET',
+    POST: 'POST',
+    PUT: 'PUT',
+    DELETE: 'DELETE',
+    PATCH: 'PATCH',
+    OPTIONS: 'OPTIONS'
   };
 
   var Events = {
-    READY_STATE_CHANGE: "readystatechange",
-    LOAD_START: "loadstart",
-    PROGRESS: "progress",
-    ABORT: "abort",
-    ERROR: "error",
-    LOAD: "load",
-    TIMEOUT: "timeout",
-    LOAD_END: "loadend"
+    READY_STATE_CHANGE: 'readystatechange',
+    LOAD_START: 'loadstart',
+    PROGRESS: 'progress',
+    ABORT: 'abort',
+    ERROR: 'error',
+    LOAD: 'load',
+    TIMEOUT: 'timeout',
+    LOAD_END: 'loadend'
   };
 
   var defaults = {
     method: Methods.GET,
     data: undefined,
     headers: {
-      Accept: "application/json",
-      "Content-Type": "application/json"
+      'Accept': 'application/json',
+      'Content-Type': 'application/json'
     },
     dump: JSON.stringify,
     load: JSON.parse,
-    xmlHttpRequest: function () {
+    xmlHttpRequest: function xmlHttpRequest() {
       return new XMLHttpRequest();
     },
-    promise: function (fn) {
+    promise: function promise(fn) {
       return new Promise(fn);
     }
   };
 
   var config = {};
 
-  var configure = function (opts) {
+  var configure = function configure(opts) {
     config = assign({}, config, opts);
   };
 
-  var promise = function (args, fn) {
+  var promise = function promise(args, fn) {
     return (args && args.promise ? args.promise : config.promise || defaults.promise)(fn);
   };
 
-  var xr = function (args) {
+  var xr = function xr(args) {
     return promise(args, function (resolve, reject) {
       var opts = assign({}, defaults, config, args);
       var xhr = opts.xmlHttpRequest();
 
-      xhr.open(opts.method, opts.params ? "" + opts.url.split("?")[0] + "?" + getParams(opts.params) : opts.url, true);
+      xhr.open(opts.method, opts.params ? '' + opts.url.split('?')[0] + '?' + getParams(opts.params) : opts.url, true);
 
       xhr.addEventListener(Events.LOAD, function () {
         return xhr.status >= 200 && xhr.status < 300 ? resolve(assign({}, res(xhr), {
@@ -115,7 +121,9 @@
         xhr.setRequestHeader(header, opts.headers[header]);
       }for (var _event in opts.events) {
         xhr.addEventListener(_event, opts.events[_event].bind(null, xhr), false);
-      }xhr.send(typeof opts.data === "object" && !opts.raw ? opts.dump(opts.data) : opts.data);
+      }var data = typeof opts.data === 'object' && !opts.raw ? opts.dump(opts.data) : opts.data;
+
+      data ? xhr.send(data) : xhr.send();
     });
   };